double a

+++ Hello.java
@@ -1,6 +1,6 @@
 public class Hello {
     public static void main(String[] args) {
-        int a =5;
+        double a =5;
         
         System.out.println("Hello, Java!!!!!!! AND GIT !!!");
